// src/app/models/vigencia.ts
export type Vigencia = {
    Id:                number;
    Nombre:            string;
    Descripcion:       string;
    Year:              number;
    Ciclo:             string;
    CodigoAbreviacion: string;
    Activo:            boolean;
    AplicacionId:      number;
    InicioVigencia:    string;
    FinVigencia:       string;
    FechaCreacion:     string;
    FechaModificacion: string;
}